+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [6.9.1] - 2022-09-26
+
 ### Added
 
 * Added `Kreait\Firebase\RemoteConfig\Template::conditionNames()` to return a list of condition names 
@@ -316,7 +318,8 @@ methods.
     * `Kreait\Firebase\Value\Uid`
     * `Kreait\Firebase\Value\Url`
 
-[Unreleased]: https://github.com/kreait/firebase-php/compare/6.9.0...6.x
+[Unreleased]: https://github.com/kreait/firebase-php/compare/6.9.1...6.x
+[6.9.1]: https://github.com/kreait/firebase-php/compare/6.9.0...6.9.1
 [6.9.0]: https://github.com/kreait/firebase-php/compare/6.8.0...6.9.0
 [6.8.0]: https://github.com/kreait/firebase-php/compare/6.7.1...6.8.0
 [6.7.1]: https://github.com/kreait/firebase-php/compare/6.7.0...6.7.1